pkg/endpoint: move UpdateProxyStatistics to proxystats.go
Signed-off by: Ian Vernon <ian@cilium.io>
