+++ pkg/endpoint/proxystats.go
@@ -18,6 +18,7 @@ import (
 	"sort"
 
 	"github.com/cilium/cilium/api/v1/models"
+	"github.com/cilium/cilium/pkg/proxy/accesslog"
 )
 
 // sortProxyStats sorts the given slice of ProxyStatistics.
@@ -50,4 +51,31 @@ func sortProxyStats(proxyStats []*models.ProxyStatistics) {
 		}
 		return false
 	})
+}
+
+// UpdateProxyStatistics updates the Endpoint's proxy  statistics to account
+// for a new observed flow with the given characteristics.
+func (e *Endpoint) UpdateProxyStatistics(l7Protocol string, port uint16, ingress, request bool, verdict accesslog.FlowVerdict) {
+	e.proxyStatisticsMutex.Lock()
+	defer e.proxyStatisticsMutex.Unlock()
+
+	proxyStats := e.getProxyStatisticsLocked(l7Protocol, port, ingress)
+
+	var stats *models.MessageForwardingStatistics
+	if request {
+		stats = proxyStats.Statistics.Requests
+	} else {
+		stats = proxyStats.Statistics.Responses
+	}
+
+	stats.Received++
+
+	switch verdict {
+	case accesslog.VerdictForwarded:
+		stats.Forwarded++
+	case accesslog.VerdictDenied:
+		stats.Denied++
+	case accesslog.VerdictError:
+		stats.Error++
+	}
 }
